OAnQuan_AI/requirements.txt
pygame-ce>=2.0.0
pytest>=7.0,<9.0
